requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.9.1
cryptography==45.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.0
django-anymail==13.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-redis==5.4.0
django-templated-mail==1.1.1
django-userforeignkey==0.5.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
model-bakery==1.20.5
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
validate_email==1.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
